requirements.txt
django
djangorestframework
django-resized
Pillow
django-cleanup
django-filter